# apps/rate-limiter-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Rate Limiter SaaS API
  description: >
    Multi-tenant distributed rate limiter. Tenants sign up, create API keys,
    and configure rate-limit rules (fixed window, sliding window, leaky
    bucket, or token bucket) per resource. Client applications then call
    `/v1/rate-limit/check` with their API key to enforce those limits.
  version: 1.0.0
servers:
  - url: http://localhost:3333
    description: Local development
...
security: []
paths:
...
  /v1/account/rules:
    get:
      tags: [Account]
      operationId: listAccountRules
      summary: List rate-limit rules for the signed-in tenant
      security: [{ SessionAuth: [] }]
      responses:
        '200':
          description: The tenant's rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RateLimitRule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Account]
      operationId: createAccountRule
      summary: Create a rule for one of the tenant's own API keys
      description: >
        `apiKeyId` must belong to the signed-in tenant (verified server-side)
        — the rest of the body is one of the four algorithm-specific shapes.
      security: [{ SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRuleInput' }
      responses:
        '201':
          description: The created rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Not signed in, or apiKeyId does not belong to this tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/account/rules/{id}:
    get:
      tags: [Account]
      operationId: getAccountRule
      summary: Get one rule
      security: [{ SessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      responses:
        '200':
          description: The rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }
    patch:
      tags: [Account]
      operationId: updateAccountRule
      summary: Update a rule's numeric parameters or active state
      description: Algorithm and resource are immutable once a rule exists.
      security: [{ SessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRuleInput' }
      responses:
        '200':
          description: The updated rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }
    delete:
      tags: [Account]
      operationId: deleteAccountRule
      summary: Deactivate a rule (soft delete)
      security: [{ SessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      responses:
        '204': { description: Deactivated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }
  /v1/account/rules/{id}/usage:
    get:
      tags: [Account]
      operationId: getAccountRuleUsage
      summary: Check current usage for a rule without consuming a request
      description: >
        Read-only — reports what `/v1/rate-limit/check` would currently
        return, computed from live Redis state, but never mutates it.
        `identifier` is required when the rule's scope is `IDENTIFIER`,
        since usage is tracked per-identifier in that case.
      security: [{ SessionAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RuleId'
        - name: identifier
          in: query
          required: false
          description: Required if the rule's scope is IDENTIFIER.
          schema: { type: string }
      responses:
        '200':
          description: Current usage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleUsage' }
        '400':
          description: Missing identifier for an IDENTIFIER-scoped rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }

  # --------------------------------------------------------------- Admin
...
  /v1/rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: List rules for the authenticated tenant
      security: [{ ApiKeyAuth: [] }]
      responses:
        '200':
          description: The tenant's rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RateLimitRule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Rules]
      operationId: createRule
      summary: Create a rule for the API key used to authenticate
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRuleInput' }
      responses:
        '201':
          description: The created rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/rules/{id}:
    get:
      tags: [Rules]
      operationId: getRule
      summary: Get one rule
      security: [{ ApiKeyAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      responses:
        '200':
          description: The rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }
    patch:
      tags: [Rules]
      operationId: updateRule
      summary: Update a rule's numeric parameters or active state
      security: [{ ApiKeyAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRuleInput' }
      responses:
        '200':
          description: The updated rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateLimitRule' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }
    delete:
      tags: [Rules]
      operationId: deleteRule
      summary: Deactivate a rule (soft delete)
      security: [{ ApiKeyAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RuleId' }]
      responses:
        '204': { description: Deactivated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/RuleNotFound' }

  # --------------------------------------------------------- Rate Limit
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: A tenant's raw API key (shown once at creation time).
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static operator token (ADMIN_TOKEN env var).
    SessionAuth:
      type: apiKey
      in: cookie
      name: rl_session
      description: httpOnly JWT session cookie set by /v1/auth/signup or /v1/auth/login.
  parameters:
    RuleId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RuleNotFound:
      description: No rule with that id on this tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
...
    RateLimitAlgorithm:
      type: string
      enum: [FIXED_WINDOW, SLIDING_WINDOW, LEAKY_BUCKET, TOKEN_BUCKET]
    RuleScope:
      type: string
      enum: [API_KEY, IDENTIFIER]
      default: API_KEY
      description: >
        API_KEY limits the whole key globally; IDENTIFIER tracks a separate
        counter per caller-supplied `identifier` within that key.
    RateLimitRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        apiKeyId: { type: string, format: uuid }
        resource: { type: string }
        algorithm: { $ref: '#/components/schemas/RateLimitAlgorithm' }
        scope: { $ref: '#/components/schemas/RuleScope' }
        limit: { type: integer }
        windowMs: { type: integer, nullable: true }
        capacity: { type: integer, nullable: true }
        refillRatePerMs: { type: number, nullable: true }
        leakRatePerMs: { type: number, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRuleInput:
      description: >
        `apiKeyId` is required on /v1/account/rules (which api key the rule
        applies to); it's implied by the authenticating key on /v1/rules.
        One of four shapes depending on `algorithm`.
      oneOf:
        - $ref: '#/components/schemas/CreateFixedOrSlidingWindowRule'
        - $ref: '#/components/schemas/CreateLeakyBucketRule'
        - $ref: '#/components/schemas/CreateTokenBucketRule'
      discriminator:
        propertyName: algorithm
        mapping:
          FIXED_WINDOW: '#/components/schemas/CreateFixedOrSlidingWindowRule'
          SLIDING_WINDOW: '#/components/schemas/CreateFixedOrSlidingWindowRule'
          LEAKY_BUCKET: '#/components/schemas/CreateLeakyBucketRule'
          TOKEN_BUCKET: '#/components/schemas/CreateTokenBucketRule'
    CreateFixedOrSlidingWindowRule:
      type: object
      required: [resource, algorithm, limit, windowMs]
      properties:
        apiKeyId: { type: string, format: uuid }
        resource: { type: string, minLength: 1, maxLength: 200 }
        algorithm:
          type: string
          enum: [FIXED_WINDOW, SLIDING_WINDOW]
        scope: { $ref: '#/components/schemas/RuleScope' }
        limit: { type: integer, minimum: 1 }
        windowMs: { type: integer, minimum: 1, description: Window length in ms }
    CreateLeakyBucketRule:
      type: object
      required: [resource, algorithm, limit, leakRatePerMs]
      properties:
        apiKeyId: { type: string, format: uuid }
        resource: { type: string, minLength: 1, maxLength: 200 }
        algorithm: { type: string, enum: [LEAKY_BUCKET] }
        scope: { $ref: '#/components/schemas/RuleScope' }
        limit: { type: integer, minimum: 1 }
        capacity: { type: integer, minimum: 1, description: Defaults to `limit` if omitted }
        leakRatePerMs: { type: number, minimum: 0, exclusiveMinimum: true, description: Units drained per ms }
    CreateTokenBucketRule:
      type: object
      required: [resource, algorithm, limit, refillRatePerMs]
      properties:
        apiKeyId: { type: string, format: uuid }
        resource: { type: string, minLength: 1, maxLength: 200 }
        algorithm: { type: string, enum: [TOKEN_BUCKET] }
        scope: { $ref: '#/components/schemas/RuleScope' }
        limit: { type: integer, minimum: 1 }
        capacity: { type: integer, minimum: 1, description: Defaults to `limit` if omitted }
        refillRatePerMs: { type: number, minimum: 0, exclusiveMinimum: true, description: Tokens added per ms }
    UpdateRuleInput:
      type: object
      description: Algorithm and resource are immutable — only these fields may change.
      properties:
        limit: { type: integer, minimum: 1 }
        windowMs: { type: integer, minimum: 1 }
        capacity: { type: integer, minimum: 1 }
        refillRatePerMs: { type: number, minimum: 0, exclusiveMinimum: true }
        leakRatePerMs: { type: number, minimum: 0, exclusiveMinimum: true }
        isActive: { type: boolean }
    RuleUsage:
      type: object
      properties:
        limit: { type: integer }
        remaining: { type: integer }
        resetMs: { type: integer, description: Milliseconds until the window/bucket fully resets }
...